feat: review Create, Update, Delete시 drinks rating update기능 추가

+++ app/reviews/application/service.py
@@ -1,9 +1,13 @@
 import time
 from typing import Union
 
+from drinks.application.dtos import AddDrinkReviewInputDto, DeleteDrinkReviewInputDto, UpdateDrinkReviewInputDto
+from drinks.application.service import DrinkApplicationService
 from reviews.application.dtos import (
     CreateReviewInputDto,
+    CreateReviewOutputDto,
     DeleteReviewInputDto,
+    DeleteReviewOutputDto,
     FindReviewInputDto,
     FindReviewOutputDto,
     FindReviewsByDrinkIdInputDto,
@@ -11,13 +15,11 @@ from reviews.application.dtos import (
     FindReviewsByUserIdInputDto,
     FindReviewsByUserIdOutputDto,
     UpdateReviewInputDto,
-    CreateReviewOutputDto,
     UpdateReviewOutputDto,
-    DeleteReviewOutputDto,
 )
 from reviews.domain.entities import Review
 from reviews.domain.repository import ReviewRepository
-from reviews.domain.value_objects import ReviewRating, UserId, ReviewId, DrinkId
+from reviews.domain.value_objects import DrinkId, ReviewId, ReviewRating, UserId
 from shared_kernel.application.dtos import FailedOutputDto
 
 
@@ -52,7 +54,9 @@ class ReviewApplicationService:
     #     except Exception as e:
     #         return FailedOutputDto.build_system_error(message=str(e))
 
-    def create_review(self, input_dto: CreateReviewInputDto) -> Union[CreateReviewOutputDto, FailedOutputDto]:
+    def create_review(
+        self, input_dto: CreateReviewInputDto, drink_application_service: DrinkApplicationService
+    ) -> Union[CreateReviewOutputDto, FailedOutputDto]:
         try:
             drink_id = DrinkId.from_str(input_dto.drink_id)
             user_id = UserId(value=input_dto.user_id)
@@ -70,6 +74,11 @@ class ReviewApplicationService:
                 updated_at=time.time(),
             )
             self._review_repository.add(review)
+            input_dto = AddDrinkReviewInputDto(drink_id=input_dto.drink_id, drink_rating=input_dto.rating)
+            drink_add_review_output_dto = drink_application_service.add_drink_review(input_dto=input_dto)
+
+            if not drink_add_review_output_dto.status:
+                return drink_add_review_output_dto
             return CreateReviewOutputDto(
                 drink_id=str(review.drink_id),
                 user_id=str(review.user_id),
@@ -82,12 +91,15 @@ class ReviewApplicationService:
         except Exception as e:
             return FailedOutputDto.build_system_error(message=str(e))
 
-    def update_review(self, input_dto: UpdateReviewInputDto) -> Union[UpdateReviewOutputDto, FailedOutputDto]:
+    def update_review(
+        self, input_dto: UpdateReviewInputDto, drink_application_service: DrinkApplicationService
+    ) -> Union[UpdateReviewOutputDto, FailedOutputDto]:
         try:
             review_id = ReviewId.from_str(input_dto.review_id)
             review = self._review_repository.find_by_review_id(review_id)
             if review is None:
                 return FailedOutputDto.build_resource_not_found_error(f"{str(input_dto.review_id)}의 리뷰를 찾을 수 없습니다.")
+            old_rating = int(review.rating)
 
             review = Review(
                 id=review.id,
@@ -99,18 +111,36 @@ class ReviewApplicationService:
                 updated_at=time.time(),
             )
             self._review_repository.update(review)
+
+            drinks_input_dto = UpdateDrinkReviewInputDto(
+                drink_id=str(review.drink_id), old_drink_rating=old_rating, new_drink_rating=input_dto.rating
+            )
+            drink_update_review_output_dto = drink_application_service.update_drink_review(input_dto=drinks_input_dto)
+
+            if not drink_update_review_output_dto.status:
+                return drink_update_review_output_dto
             return UpdateReviewOutputDto()
 
         except Exception as e:
             return FailedOutputDto.build_system_error(message=str(e))
 
-    def delete_review(self, input_dto: DeleteReviewInputDto) -> Union[DeleteReviewOutputDto, FailedOutputDto]:
+    def delete_review(
+        self, input_dto: DeleteReviewInputDto, drink_application_service: DrinkApplicationService
+    ) -> Union[DeleteReviewOutputDto, FailedOutputDto]:
         try:
             review_id = ReviewId.from_str(input_dto.review_id)
             review = self._review_repository.find_by_review_id(review_id)
             if review is None:
                 return FailedOutputDto.build_resource_not_found_error(f"{str(input_dto.review_id)}의 리뷰를 찾을 수 없습니다.")
+
             self._review_repository.delete_by_review_id(review_id)
+            drinks_input_dto = DeleteDrinkReviewInputDto(
+                drink_id=str(review.drink_id), drink_rating=int(review.rating)
+            )
+            drink_delete_review_output_dto = drink_application_service.delete_drink_review(input_dto=drinks_input_dto)
+
+            if not drink_delete_review_output_dto.status:
+                return drink_delete_review_output_dto
             return DeleteReviewOutputDto()
 
         except Exception as e:
